Update views.py
This code is for a basic blog. The homepage shows all posts. You can click a post to see its details. If you're logged in, you can create a new post. New users can register using a signup form, and after signing up, they are logged in automatically. Only logged-in users are allowed to make posts.